Splitt ut BehandlingSendtTilTotrinnskontrollModal i egen fil
